@@ -1,6 +1,6 @@
 {
   "name": "webpack-dev-server",
-  "version": "4.0.0-rc.0",
+  "version": "4.0.0",
   "description": "Serves a webpack app. Updates the browser on changes.",
   "bin": "bin/webpack-dev-server.js",
   "main": "lib/Server.js",
@@ -21,13 +21,14 @@
     "fix:js": "npm run lint:js -- --fix",
     "fix": "npm-run-all fix:js fmt",
     "commitlint": "commitlint --from=master",
+    "build": "npm-run-all build:client",
+    "build:client": "rimraf ./client/* && babel client-src/ --out-dir client/ --ignore \"client-src/webpack.config.js\" --ignore \"client-src/modules\" && webpack --config client-src/webpack.config.js",
     "test:only": "jest",
     "test:coverage": "npm run test:only -- --coverage",
     "test:watch": "npm run test:coverage --watch",
     "test": "npm run test:coverage",
     "pretest": "npm run lint",
-    "prepare": "npm run build:client && husky install",
-    "build:client": "rimraf ./client/* && babel client-src/ --out-dir client/ --ignore \"client-src/webpack.config.js\" --ignore \"client-src/modules\" && webpack --config client-src/webpack.config.js",
+    "prepare": "husky install && npm run build",
     "release": "standard-version"
   },
   "dependencies": {
@@ -40,11 +41,10 @@
     "del": "^6.0.0",
     "express": "^4.17.1",
     "graceful-fs": "^4.2.6",
+    "html-entities": "^2.3.2",
     "http-proxy-middleware": "^2.0.0",
     "internal-ip": "^6.2.0",
     "ipaddr.js": "^2.0.1",
-    "is-absolute-url": "^3.0.3",
-    "killable": "^1.0.1",
     "open": "^8.0.9",
     "p-retry": "^4.5.0",
     "portfinder": "^1.0.28",
@@ -56,7 +56,7 @@
     "strip-ansi": "^7.0.0",
     "url": "^0.11.0",
     "webpack-dev-middleware": "^5.0.0",
-    "ws": "^7.5.3"
+    "ws": "^8.1.0"
   },
   "devDependencies": {
     "@babel/cli": "^7.14.5",
@@ -66,8 +66,8 @@
     "@babel/plugin-transform-runtime": "^7.14.5",
     "@babel/preset-env": "^7.14.5",
     "@babel/runtime": "^7.14.5",
-    "@commitlint/cli": "^12.1.3",
-    "@commitlint/config-conventional": "^12.1.4",
+    "@commitlint/cli": "^13.1.0",
+    "@commitlint/config-conventional": "^13.1.0",
     "acorn": "^8.2.4",
     "babel-jest": "^27.0.2",
     "babel-loader": "^8.2.2",
@@ -88,7 +88,7 @@
     "less": "^4.1.1",
     "less-loader": "^7.3.0",
     "lint-staged": "^11.0.0",
-    "marked": "^2.0.3",
+    "marked": "^3.0.0",
     "memfs": "^3.2.2",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.3.1",
@@ -103,12 +103,12 @@
     "tcp-port-used": "^1.0.2",
     "typescript": "^4.2.4",
     "url-loader": "^4.1.1",
-    "webpack": "^5.44.0",
+    "webpack": "^5.50.0",
     "webpack-cli": "^4.7.2",
     "webpack-merge": "^5.8.0"
   },
   "peerDependencies": {
-    "webpack": "^4.0.0 || ^5.0.0"
+    "webpack": "^4.37.0 || ^5.0.0"
   },
   "peerDependenciesMeta": {
     "webpack-cli": {
