@@ -1,7 +1,8 @@
 // The error overlay is inspired (and mostly copied) from Create React App (https://github.com/facebookincubator/create-react-app)
 // They, in turn, got inspired by webpack-hot-middleware (https://github.com/glenjamin/webpack-hot-middleware).
 
-const ansiHTML = require("ansi-html");
+import ansiHTML from "ansi-html";
+import { encode } from "html-entities";
 
 const colors = {
   reset: ["transparent", "transparent"],
@@ -135,14 +136,15 @@ function show(messages, type) {
 
       // Make it look similar to our terminal.
       const errorMessage = message.message || messages[0];
-      const text = ansiHTML(errorMessage);
-      const messageTextNode = document.createTextNode(text);
+      const text = ansiHTML(encode(errorMessage));
+      const messageTextNode = document.createElement("div");
+
+      messageTextNode.innerHTML = text;
 
       entryElement.appendChild(typeElement);
       entryElement.appendChild(document.createElement("br"));
       entryElement.appendChild(document.createElement("br"));
       entryElement.appendChild(messageTextNode);
-      entryElement.appendChild(document.createElement("br"));
       entryElement.appendChild(document.createElement("br"));
       entryElement.appendChild(document.createElement("br"));
 
