@@ -3,10 +3,10 @@
 const os = require("os");
 const path = require("path");
 const url = require("url");
+const util = require("util");
 const fs = require("graceful-fs");
 const ipaddr = require("ipaddr.js");
 const internalIp = require("internal-ip");
-const killable = require("killable");
 const express = require("express");
 const { validate } = require("schema-utils");
 const schema = require("./options.json");
@@ -19,7 +19,13 @@ class Server {
   constructor(options = {}, compiler) {
     // TODO: remove this after plugin support is published
     if (options.hooks) {
-      [options, compiler] = [compiler, options];
+      util.deprecate(
+        () => {},
+        "Using 'compiler' as the first argument is deprecated. Please use 'options' as the first argument and 'compiler' as the second argument.",
+        "DEP_WEBPACK_DEV_SERVER_CONSTRUCTOR"
+      )();
+
+      [options = {}, compiler] = [compiler, options];
     }
 
     validate(schema, options, "webpack Dev Server");
@@ -28,7 +34,7 @@ class Server {
     this.staticWatchers = [];
     // Keep track of websocket proxies for external websocket upgrade.
     this.webSocketProxies = [];
-
+    this.sockets = [];
     this.compiler = compiler;
   }
 
@@ -43,46 +49,48 @@ class Server {
     };
   }
 
-  static getHostname(hostname) {
+  // eslint-disable-next-line class-methods-use-this
+  static isAbsoluteURL(URL) {
+    // Don't match Windows paths `c:\`
+    if (/^[a-zA-Z]:\\/.test(URL)) {
+      return false;
+    }
+
+    // Scheme: https://tools.ietf.org/html/rfc3986#section-3.1
+    // Absolute URL: https://tools.ietf.org/html/rfc3986#section-4.3
+    return /^[a-zA-Z][a-zA-Z\d+\-.]*:/.test(URL);
+  }
+
+  static async getHostname(hostname) {
     if (hostname === "local-ip") {
-      return internalIp.v4.sync() || internalIp.v6.sync() || "0.0.0.0";
+      return (await internalIp.v4()) || (await internalIp.v6()) || "0.0.0.0";
     } else if (hostname === "local-ipv4") {
-      return internalIp.v4.sync() || "0.0.0.0";
+      return (await internalIp.v4()) || "0.0.0.0";
     } else if (hostname === "local-ipv6") {
-      return internalIp.v6.sync() || "::";
+      return (await internalIp.v6()) || "::";
     }
 
     return hostname;
   }
 
-  static getFreePort(port) {
+  static async getFreePort(port) {
+    if (port && port !== "auto") {
+      return port;
+    }
+
     const pRetry = require("p-retry");
     const portfinder = require("portfinder");
 
-    if (port && port !== "auto") {
-      return Promise.resolve(port);
-    }
-
-    function runPortFinder() {
-      return new Promise((resolve, reject) => {
-        // Default port
-        portfinder.basePort = process.env.WEBPACK_DEV_SERVER_BASE_PORT || 8080;
-        portfinder.getPort((error, foundPort) => {
-          if (error) {
-            return reject(error);
-          }
-
-          return resolve(foundPort);
-        });
-      });
-    }
+    portfinder.basePort = process.env.WEBPACK_DEV_SERVER_BASE_PORT || 8080;
 
     // Try to find unused port and listen on it for 3 times,
     // if port is not specified in options.
     const defaultPortRetry =
       parseInt(process.env.WEBPACK_DEV_SERVER_PORT_RETRY, 10) || 3;
 
-    return pRetry(runPortFinder, { retries: defaultPortRetry });
+    return pRetry(() => portfinder.getPortPromise(), {
+      retries: defaultPortRetry,
+    });
   }
 
   static findCacheDir() {
@@ -118,25 +126,284 @@ class Server {
     return path.resolve(dir, "node_modules/.cache/webpack-dev-server");
   }
 
-  getCompilerConfigArray() {
-    const compilers = this.compiler.compilers
-      ? this.compiler.compilers
-      : [this.compiler];
+  addAdditionalEntries(compiler) {
+    const additionalEntries = [];
 
-    return compilers.map((compiler) => compiler.options);
+    const isWebTarget = (compilerOptions) =>
+      compilerOptions.externalsPresets
+        ? compilerOptions.externalsPresets.web
+        : [
+            "web",
+            "webworker",
+            "electron-preload",
+            "electron-renderer",
+            "node-webkit",
+            // eslint-disable-next-line no-undefined
+            undefined,
+            null,
+          ].includes(compilerOptions.target);
+
+    // TODO maybe empty empty client
+    if (this.options.client && isWebTarget(compiler.options)) {
+      const getWebSocketURL = () => {
+        const searchParams = new URLSearchParams();
+
+        /** @type {"ws:" | "wss:" | "http:" | "https:" | "auto:"} */
+        let protocol;
+
+        // We are proxying dev server and need to specify custom `hostname`
+        if (typeof this.options.client.webSocketURL.protocol !== "undefined") {
+          protocol = this.options.client.webSocketURL.protocol;
+        } else {
+          protocol = this.options.https ? "wss:" : "ws:";
+        }
+
+        searchParams.set("protocol", protocol);
+
+        if (typeof this.options.client.webSocketURL.username !== "undefined") {
+          searchParams.set(
+            "username",
+            this.options.client.webSocketURL.username
+          );
+        }
+
+        if (typeof this.options.client.webSocketURL.password !== "undefined") {
+          searchParams.set(
+            "password",
+            this.options.client.webSocketURL.password
+          );
+        }
+
+        /** @type {string} */
+        let hostname;
+
+        // SockJS is not supported server mode, so `hostname` and `port` can't specified, let's ignore them
+        // TODO show warning about this
+        const isSockJSType = this.options.webSocketServer.type === "sockjs";
+
+        // We are proxying dev server and need to specify custom `hostname`
+        if (typeof this.options.client.webSocketURL.hostname !== "undefined") {
+          hostname = this.options.client.webSocketURL.hostname;
+        }
+        // Web socket server works on custom `hostname`, only for `ws` because `sock-js` is not support custom `hostname`
+        else if (
+          typeof this.options.webSocketServer.options.host !== "undefined" &&
+          !isSockJSType
+        ) {
+          hostname = this.options.webSocketServer.options.host;
+        }
+        // The `host` option is specified
+        else if (typeof this.options.host !== "undefined") {
+          hostname = this.options.host;
+        }
+        // The `port` option is not specified
+        else {
+          hostname = "0.0.0.0";
+        }
+
+        searchParams.set("hostname", hostname);
+
+        /** @type {number | string} */
+        let port;
+
+        // We are proxying dev server and need to specify custom `port`
+        if (typeof this.options.client.webSocketURL.port !== "undefined") {
+          port = this.options.client.webSocketURL.port;
+        }
+        // Web socket server works on custom `port`, only for `ws` because `sock-js` is not support custom `port`
+        else if (
+          typeof this.options.webSocketServer.options.port !== "undefined" &&
+          !isSockJSType
+        ) {
+          port = this.options.webSocketServer.options.port;
+        }
+        // The `port` option is specified
+        else if (typeof this.options.port === "number") {
+          port = this.options.port;
+        }
+        // The `port` option is specified using `string`
+        else if (
+          typeof this.options.port === "string" &&
+          this.options.port !== "auto"
+        ) {
+          port = Number(this.options.port);
+        }
+        // The `port` option is not specified or set to `auto`
+        else {
+          port = "0";
+        }
+
+        searchParams.set("port", String(port));
+
+        /** @type {string} */
+        let pathname = "";
+
+        // We are proxying dev server and need to specify custom `pathname`
+        if (typeof this.options.client.webSocketURL.pathname !== "undefined") {
+          pathname = this.options.client.webSocketURL.pathname;
+        }
+        // Web socket server works on custom `path`
+        else if (
+          typeof this.options.webSocketServer.options.prefix !== "undefined" ||
+          typeof this.options.webSocketServer.options.path !== "undefined"
+        ) {
+          pathname =
+            this.options.webSocketServer.options.prefix ||
+            this.options.webSocketServer.options.path;
+        }
+
+        searchParams.set("pathname", pathname);
+
+        if (typeof this.options.client.logging !== "undefined") {
+          searchParams.set("logging", this.options.client.logging);
+        }
+
+        return searchParams.toString();
+      };
+      const webSocketURL = this.options.webSocketServer
+        ? getWebSocketURL()
+        : "";
+
+      additionalEntries.push(
+        `${require.resolve("../client/index.js")}?${webSocketURL}`
+      );
+    }
+
+    if (this.options.hot) {
+      let hotEntry;
+
+      if (this.options.hot === "only") {
+        hotEntry = require.resolve("webpack/hot/only-dev-server");
+      } else if (this.options.hot) {
+        hotEntry = require.resolve("webpack/hot/dev-server");
+      }
+
+      additionalEntries.push(hotEntry);
+    }
+
+    const webpack = compiler.webpack || require("webpack");
+
+    // use a hook to add entries if available
+    if (typeof webpack.EntryPlugin !== "undefined") {
+      for (const additionalEntry of additionalEntries) {
+        new webpack.EntryPlugin(compiler.context, additionalEntry, {
+          // eslint-disable-next-line no-undefined
+          name: undefined,
+        }).apply(compiler);
+      }
+    }
+    // TODO remove after drop webpack v4 support
+    else {
+      /**
+       * prependEntry Method for webpack 4
+       * @param {Entry} originalEntry
+       * @param {Entry} newAdditionalEntries
+       * @returns {Entry}
+       */
+      const prependEntry = (originalEntry, newAdditionalEntries) => {
+        if (typeof originalEntry === "function") {
+          return () =>
+            Promise.resolve(originalEntry()).then((entry) =>
+              prependEntry(entry, newAdditionalEntries)
+            );
+        }
+
+        if (
+          typeof originalEntry === "object" &&
+          !Array.isArray(originalEntry)
+        ) {
+          /** @type {Object<string,string>} */
+          const clone = {};
+
+          Object.keys(originalEntry).forEach((key) => {
+            // entry[key] should be a string here
+            const entryDescription = originalEntry[key];
+
+            clone[key] = prependEntry(entryDescription, newAdditionalEntries);
+          });
+
+          return clone;
+        }
+
+        // in this case, entry is a string or an array.
+        // make sure that we do not add duplicates.
+        /** @type {Entry} */
+        const entriesClone = additionalEntries.slice(0);
+
+        [].concat(originalEntry).forEach((newEntry) => {
+          if (!entriesClone.includes(newEntry)) {
+            entriesClone.push(newEntry);
+          }
+        });
+
+        return entriesClone;
+      };
+
+      compiler.options.entry = prependEntry(
+        compiler.options.entry || "./src",
+        additionalEntries
+      );
+      compiler.hooks.entryOption.call(
+        compiler.options.context,
+        compiler.options.entry
+      );
+    }
+  }
+
+  getCompilerOptions() {
+    if (typeof this.compiler.compilers !== "undefined") {
+      if (this.compiler.compilers.length === 1) {
+        return this.compiler.compilers[0].options;
+      }
+
+      // Configuration with the `devServer` options
+      const compilerWithDevServer = this.compiler.compilers.find(
+        (config) => config.options.devServer
+      );
+
+      if (compilerWithDevServer) {
+        return compilerWithDevServer.options;
+      }
+
+      // Configuration with `web` preset
+      const compilerWithWebPreset = this.compiler.compilers.find(
+        (config) =>
+          (config.options.externalsPresets &&
+            config.options.externalsPresets.web) ||
+          [
+            "web",
+            "webworker",
+            "electron-preload",
+            "electron-renderer",
+            "node-webkit",
+            // eslint-disable-next-line no-undefined
+            undefined,
+            null,
+          ].includes(config.options.target)
+      );
+
+      if (compilerWithWebPreset) {
+        return compilerWithWebPreset.options;
+      }
+
+      // Fallback
+      return this.compiler.compilers[0].options;
+    }
+
+    return this.compiler.options;
   }
 
   // eslint-disable-next-line class-methods-use-this
-  normalizeOptions(options) {
-    // TODO: improve this to not use .find for compiler watchOptions
-    const configArray = this.getCompilerConfigArray(this.compiler);
-    const watchOptionsConfig = configArray.find(
-      (config) => config.watch !== false && config.watchOptions
-    );
-    const watchOptions = watchOptionsConfig
-      ? watchOptionsConfig.watchOptions
-      : {};
+  async normalizeOptions() {
+    const { options } = this;
 
+    if (!this.logger) {
+      this.logger = this.compiler.getInfrastructureLogger("webpack-dev-server");
+    }
+
+    const compilerOptions = this.getCompilerOptions();
+    // TODO remove `{}` after drop webpack v4 support
+    const watchOptions = compilerOptions.watchOptions || {};
     const defaultOptionsForStatic = {
       directory: path.join(process.cwd(), "public"),
       staticOptions: {},
@@ -147,17 +414,23 @@ class Server {
     };
 
     if (typeof options.allowedHosts === "undefined") {
-      // allowedHosts allows some default hosts picked from
-      // `options.host` or `webSocketURL.hostname` and `localhost`
+      // AllowedHosts allows some default hosts picked from `options.host` or `webSocketURL.hostname` and `localhost`
       options.allowedHosts = "auto";
     }
-    if (
+    // We store allowedHosts as array when supplied as string
+    else if (
       typeof options.allowedHosts === "string" &&
       options.allowedHosts !== "auto" &&
       options.allowedHosts !== "all"
     ) {
-      // we store allowedHosts as array when supplied as string
       options.allowedHosts = [options.allowedHosts];
+    }
+    // CLI pass options as array, we should normalize them
+    else if (
+      Array.isArray(this.options.allowedHosts) &&
+      this.options.allowedHosts.includes("all")
+    ) {
+      options.allowedHosts = "all";
     }
 
     if (typeof options.bonjour === "undefined") {
@@ -202,6 +475,13 @@ class Server {
           warnings: true,
           ...options.client.overlay,
         };
+      }
+
+      // Respect infrastructureLogging.level
+      if (typeof options.client.logging === "undefined") {
+        options.client.logging = compilerOptions.infrastructureLogging
+          ? compilerOptions.infrastructureLogging.level
+          : "info";
       }
     }
 
@@ -265,11 +545,18 @@ class Server {
       if (!options.https.key || !options.https.cert) {
         const certificateDir = Server.findCacheDir();
         const certificatePath = path.join(certificateDir, "server.pem");
-        let certificateExists = fs.existsSync(certificatePath);
+        let certificateExists;
+
+        try {
+          const certificate = await fs.promises.stat(certificatePath);
+          certificateExists = certificate.isFile();
+        } catch {
+          certificateExists = false;
+        }
 
         if (certificateExists) {
           const certificateTtl = 1000 * 60 * 60 * 24;
-          const certificateStat = fs.statSync(certificatePath);
+          const certificateStat = await fs.promises.stat(certificatePath);
 
           const now = new Date();
 
@@ -278,17 +565,17 @@ class Server {
             const del = require("del");
 
             this.logger.info(
-              "SSL Certificate is more than 30 days old. Removing."
+              "SSL Certificate is more than 30 days old. Removing..."
             );
 
-            del.sync([certificatePath], { force: true });
+            await del([certificatePath], { force: true });
 
             certificateExists = false;
           }
         }
 
         if (!certificateExists) {
-          this.logger.info("Generating SSL Certificate");
+          this.logger.info("Generating SSL Certificate...");
 
           const selfsigned = require("selfsigned");
           const attributes = [{ name: "commonName", value: "localhost" }];
@@ -297,10 +584,10 @@ class Server {
             days: 30,
             keySize: 2048,
             extensions: [
-              // {
-              //   name: 'basicConstraints',
-              //   cA: true,
-              // },
+              {
+                name: "basicConstraints",
+                cA: true,
+              },
               {
                 name: "keyUsage",
                 keyCertSign: true,
@@ -354,13 +641,20 @@ class Server {
             ],
           });
 
-          fs.mkdirSync(certificateDir, { recursive: true });
-          fs.writeFileSync(certificatePath, pems.private + pems.cert, {
-            encoding: "utf8",
-          });
+          await fs.promises.mkdir(certificateDir, { recursive: true });
+
+          await fs.promises.writeFile(
+            certificatePath,
+            pems.private + pems.cert,
+            {
+              encoding: "utf8",
+            }
+          );
         }
 
-        fakeCert = fs.readFileSync(certificatePath);
+        fakeCert = await fs.promises.readFile(certificatePath);
+
+        this.logger.info(`SSL certificate: ${certificatePath}`);
       }
 
       options.https.key = options.https.key || fakeCert;
@@ -377,6 +671,9 @@ class Server {
 
     options.liveReload =
       typeof options.liveReload !== "undefined" ? options.liveReload : true;
+
+    options.magicHtml =
+      typeof options.magicHtml !== "undefined" ? options.magicHtml : true;
 
     // https://github.com/webpack/webpack-dev-server/issues/1990
     const defaultOpenOptions = { wait: false };
@@ -441,8 +738,12 @@ class Server {
      * }
      */
     if (typeof options.proxy !== "undefined") {
+      // TODO remove in the next major release, only accept `Array`
       if (!Array.isArray(options.proxy)) {
-        if (Object.prototype.hasOwnProperty.call(options.proxy, "target")) {
+        if (
+          Object.prototype.hasOwnProperty.call(options.proxy, "target") ||
+          Object.prototype.hasOwnProperty.call(options.proxy, "router")
+        ) {
           options.proxy = [options.proxy];
         } else {
           options.proxy = Object.keys(options.proxy).map((context) => {
@@ -462,39 +763,42 @@ class Server {
               proxyOptions.context = correctedContext;
             }
 
-            const getLogLevelForProxy = (level) => {
-              if (level === "none") {
-                return "silent";
-              }
-
-              if (level === "log") {
-                return "info";
-              }
-
-              if (level === "verbose") {
-                return "debug";
-              }
-
-              return level;
-            };
-
-            const configWithDevServer =
-              configArray.find((config) => config.devServer) || configArray[0];
-
-            if (typeof proxyOptions.logLevel === "undefined") {
-              proxyOptions.logLevel = getLogLevelForProxy(
-                configWithDevServer.infrastructureLogging.level
-              );
-            }
-
-            if (typeof proxyOptions.logProvider === "undefined") {
-              proxyOptions.logProvider = () => this.logger;
-            }
-
             return proxyOptions;
           });
         }
       }
+
+      options.proxy = options.proxy.map((item) => {
+        const getLogLevelForProxy = (level) => {
+          if (level === "none") {
+            return "silent";
+          }
+
+          if (level === "log") {
+            return "info";
+          }
+
+          if (level === "verbose") {
+            return "debug";
+          }
+
+          return level;
+        };
+
+        if (typeof item.logLevel === "undefined") {
+          item.logLevel = getLogLevelForProxy(
+            compilerOptions.infrastructureLogging
+              ? compilerOptions.infrastructureLogging.level
+              : "info"
+          );
+        }
+
+        if (typeof item.logProvider === "undefined") {
+          item.logProvider = () => this.logger;
+        }
+
+        return item;
+      });
     }
 
     if (typeof options.setupExitSignals === "undefined") {
@@ -522,10 +826,8 @@ class Server {
     }
 
     if (options.static) {
-      const isAbsoluteUrl = require("is-absolute-url");
-
       options.static.forEach((staticOption) => {
-        if (isAbsoluteUrl(staticOption.directory)) {
+        if (Server.isAbsoluteURL(staticOption.directory)) {
           throw new Error("Using a URL as static.directory is not supported");
         }
 
@@ -609,55 +911,103 @@ class Server {
     }
   }
 
-  initialize() {
-    this.applyDevServerPlugin();
+  getClientTransport() {
+    let ClientImplementation;
+    let clientImplementationFound = true;
 
-    if (this.options.client && this.options.client.progress) {
-      this.setupProgressPlugin();
+    const isKnownWebSocketServerImplementation =
+      this.options.webSocketServer &&
+      typeof this.options.webSocketServer.type === "string" &&
+      (this.options.webSocketServer.type === "ws" ||
+        this.options.webSocketServer.type === "sockjs");
+
+    let clientTransport;
+
+    if (this.options.client) {
+      if (typeof this.options.client.webSocketTransport !== "undefined") {
+        clientTransport = this.options.client.webSocketTransport;
+      } else if (isKnownWebSocketServerImplementation) {
+        clientTransport = this.options.webSocketServer.type;
+      } else {
+        clientTransport = "ws";
+      }
+    } else {
+      clientTransport = "ws";
     }
 
-    this.setupHooks();
-    this.setupApp();
-    this.setupHostHeaderCheck();
-    this.setupDevMiddleware();
-    // Should be after `webpack-dev-middleware`, otherwise other middlewares might rewrite response
-    this.setupBuiltInRoutes();
-    this.setupWatchFiles();
-    this.setupFeatures();
-    this.createServer();
-
-    killable(this.server);
-
-    if (this.options.setupExitSignals) {
-      const signals = ["SIGINT", "SIGTERM"];
-
-      signals.forEach((signal) => {
-        process.on(signal, () => {
-          this.close(() => {
-            // eslint-disable-next-line no-process-exit
-            process.exit();
-          });
-        });
-      });
+    switch (typeof clientTransport) {
+      case "string":
+        // could be 'sockjs', 'ws', or a path that should be required
+        if (clientTransport === "sockjs") {
+          ClientImplementation = require.resolve(
+            "../client/clients/SockJSClient"
+          );
+        } else if (clientTransport === "ws") {
+          ClientImplementation = require.resolve(
+            "../client/clients/WebSocketClient"
+          );
+        } else {
+          try {
+            // eslint-disable-next-line import/no-dynamic-require
+            ClientImplementation = require.resolve(clientTransport);
+          } catch (e) {
+            clientImplementationFound = false;
+          }
+        }
+        break;
+      default:
+        clientImplementationFound = false;
     }
 
-    // Proxy WebSocket without the initial http request
-    // https://github.com/chimurai/http-proxy-middleware#external-websocket-upgrade
-    // eslint-disable-next-line func-names
-    this.webSocketProxies.forEach(function (webSocketProxy) {
-      this.server.on("upgrade", webSocketProxy.upgrade);
-    }, this);
+    if (!clientImplementationFound) {
+      throw new Error(
+        `${
+          !isKnownWebSocketServerImplementation
+            ? "When you use custom web socket implementation you must explicitly specify client.webSocketTransport. "
+            : ""
+        }client.webSocketTransport must be a string denoting a default implementation (e.g. 'sockjs', 'ws') or a full path to a JS file via require.resolve(...) which exports a class `
+      );
+    }
+
+    return ClientImplementation;
   }
 
-  applyDevServerPlugin() {
-    const DevServerPlugin = require("./utils/DevServerPlugin");
+  getServerTransport() {
+    let implementation;
+    let implementationFound = true;
 
-    const compilers = this.compiler.compilers || [this.compiler];
+    switch (typeof this.options.webSocketServer.type) {
+      case "string":
+        // Could be 'sockjs', in the future 'ws', or a path that should be required
+        if (this.options.webSocketServer.type === "sockjs") {
+          implementation = require("./servers/SockJSServer");
+        } else if (this.options.webSocketServer.type === "ws") {
+          implementation = require("./servers/WebsocketServer");
+        } else {
+          try {
+            // eslint-disable-next-line import/no-dynamic-require
+            implementation = require(this.options.webSocketServer.type);
+          } catch (error) {
+            implementationFound = false;
+          }
+        }
+        break;
+      case "function":
+        implementation = this.options.webSocketServer.type;
+        break;
+      default:
+        implementationFound = false;
+    }
 
-    // eslint-disable-next-line no-shadow
-    compilers.forEach((compiler) => {
-      new DevServerPlugin(this.options).apply(compiler);
-    });
+    if (!implementationFound) {
+      throw new Error(
+        "webSocketServer (webSocketServer.type) must be a string denoting a default implementation (e.g. 'ws', 'sockjs'), a full path to " +
+          "a JS file which exports a class extending BaseServer (webpack-dev-server/lib/servers/BaseServer.js) " +
+          "via require.resolve(...), or the class itself which extends BaseServer"
+      );
+    }
+
+    return implementation;
   }
 
   setupProgressPlugin() {
@@ -688,10 +1038,91 @@ class Server {
     }).apply(this.compiler);
   }
 
+  async initialize() {
+    const compilers = this.compiler.compilers || [this.compiler];
+
+    // eslint-disable-next-line no-shadow
+    compilers.forEach((compiler) => {
+      this.addAdditionalEntries(compiler);
+
+      const webpack = compiler.webpack || require("webpack");
+
+      const providePlugin = new webpack.ProvidePlugin({
+        __webpack_dev_server_client__: this.getClientTransport(),
+      });
+
+      providePlugin.apply(compiler);
+
+      // TODO remove after drop webpack v4 support
+      compiler.options.plugins = compiler.options.plugins || [];
+
+      if (this.options.hot) {
+        const HMRPluginExists = compiler.options.plugins.find(
+          (p) => p.constructor === webpack.HotModuleReplacementPlugin
+        );
+
+        if (HMRPluginExists) {
+          this.logger.warn(
+            `"hot: true" automatically applies HMR plugin, you don't have to add it manually to your webpack configuration.`
+          );
+        } else {
+          // apply the HMR plugin
+          const plugin = new webpack.HotModuleReplacementPlugin();
+          plugin.apply(compiler);
+        }
+      }
+    });
+
+    if (this.options.client && this.options.client.progress) {
+      this.setupProgressPlugin();
+    }
+
+    this.setupHooks();
+    this.setupApp();
+    this.setupHostHeaderCheck();
+    this.setupDevMiddleware();
+    // Should be after `webpack-dev-middleware`, otherwise other middlewares might rewrite response
+    this.setupBuiltInRoutes();
+    this.setupWatchFiles();
+    this.setupFeatures();
+    this.createServer();
+
+    if (this.options.setupExitSignals) {
+      const signals = ["SIGINT", "SIGTERM"];
+
+      signals.forEach((signal) => {
+        process.on(signal, () => {
+          this.stopCallback(() => {
+            // eslint-disable-next-line no-process-exit
+            process.exit();
+          });
+        });
+      });
+    }
+
+    // Proxy WebSocket without the initial http request
+    // https://github.com/chimurai/http-proxy-middleware#external-websocket-upgrade
+    // eslint-disable-next-line func-names
+    this.webSocketProxies.forEach(function (webSocketProxy) {
+      this.server.on("upgrade", webSocketProxy.upgrade);
+    }, this);
+  }
+
   setupApp() {
     // Init express server
     // eslint-disable-next-line new-cap
     this.app = new express();
+  }
+
+  getStats(statsObj) {
+    const stats = Server.DEFAULT_STATS;
+    const compilerOptions = this.getCompilerOptions();
+
+    if (compilerOptions.stats && compilerOptions.stats.warningsFilter) {
+      stats.warningsFilter = compilerOptions.stats.warningsFilter;
+    }
+
+    return statsObj.toJson(stats);
   }
 
   setupHooks() {
@@ -743,7 +1174,7 @@ class Server {
     app.get("/__webpack_dev_server__/sockjs.bundle.js", (req, res) => {
       res.setHeader("Content-Type", "application/javascript");
 
-      const { createReadStream } = require("graceful-fs");
+      const { createReadStream } = fs;
       const clientPath = path.join(__dirname, "..", "client");
 
       createReadStream(
@@ -812,26 +1243,30 @@ class Server {
     const { createProxyMiddleware } = require("http-proxy-middleware");
 
     const getProxyMiddleware = (proxyConfig) => {
-      const context = proxyConfig.context || proxyConfig.path;
-
-      // It is possible to use the `bypass` method without a `target`.
+      // It is possible to use the `bypass` method without a `target` or `router`.
       // However, the proxy middleware has no use in this case, and will fail to instantiate.
       if (proxyConfig.target) {
+        const context = proxyConfig.context || proxyConfig.path;
+
         return createProxyMiddleware(context, proxyConfig);
+      }
+
+      if (proxyConfig.router) {
+        return createProxyMiddleware(proxyConfig);
       }
     };
     /**
      * Assume a proxy configuration specified as:
      * proxy: [
      *   {
-     *     context: ...,
-     *     ...options...
+     *     context: "value",
+     *     ...options,
      *   },
      *   // or:
      *   function() {
      *     return {
-     *       context: ...,
-     *       ...options...
+     *       context: "context",
+     *       ...options,
      *     };
      *   }
      * ]
@@ -863,6 +1298,7 @@ class Server {
         // - Check if we have a bypass function defined
         // - In case the bypass function is defined we'll retrieve the
         // bypassUrl from it otherwise bypassUrl would be null
+        // TODO remove in the next major in favor `context` and `router` options
         const isByPassFuncDefined = typeof proxyConfig.bypass === "function";
         const bypassUrl = isByPassFuncDefined
           ? await proxyConfig.bypass(req, res, proxyConfig)
@@ -1060,7 +1496,9 @@ class Server {
       runnableFeatures.push("staticServeIndex", "staticWatch");
     }
 
-    runnableFeatures.push("magicHtml");
+    if (this.options.magicHtml) {
+      runnableFeatures.push("magicHtml");
+    }
 
     if (this.options.onAfterSetupMiddleware) {
       runnableFeatures.push("onAfterSetupMiddleware");
@@ -1072,9 +1510,6 @@ class Server {
   }
 
   createServer() {
-    const https = require("https");
-    const http = require("http");
-
     if (this.options.https) {
       if (this.options.http2) {
         // TODO: we need to replace spdy with http2 which is an internal module
@@ -1088,57 +1523,33 @@ class Server {
           this.app
         );
       } else {
+        const https = require("https");
+
         this.server = https.createServer(this.options.https, this.app);
       }
     } else {
+      const http = require("http");
+
       this.server = http.createServer(this.app);
     }
+
+    this.server.on("connection", (socket) => {
+      // Add socket to list
+      this.sockets.push(socket);
+
+      socket.once("close", () => {
+        // Remove socket from list
+        this.sockets.splice(this.sockets.indexOf(socket), 1);
+      });
+    });
 
     this.server.on("error", (error) => {
       throw error;
     });
   }
 
-  getWebSocketServerImplementation() {
-    let implementation;
-    let implementationFound = true;
-
-    switch (typeof this.options.webSocketServer.type) {
-      case "string":
-        // Could be 'sockjs', in the future 'ws', or a path that should be required
-        if (this.options.webSocketServer.type === "sockjs") {
-          implementation = require("./servers/SockJSServer");
-        } else if (this.options.webSocketServer.type === "ws") {
-          implementation = require("./servers/WebsocketServer");
-        } else {
-          try {
-            // eslint-disable-next-line import/no-dynamic-require
-            implementation = require(this.options.webSocketServer.type);
-          } catch (error) {
-            implementationFound = false;
-          }
-        }
-        break;
-      case "function":
-        implementation = this.options.webSocketServer.type;
-        break;
-      default:
-        implementationFound = false;
-    }
-
-    if (!implementationFound) {
-      throw new Error(
-        "webSocketServer (webSocketServer.type) must be a string denoting a default implementation (e.g. 'ws', 'sockjs'), a full path to " +
-          "a JS file which exports a class extending BaseServer (webpack-dev-server/lib/servers/BaseServer.js) " +
-          "via require.resolve(...), or the class itself which extends BaseServer"
-      );
-    }
-
-    return implementation;
-  }
-
   createWebSocketServer() {
-    this.webSocketServer = new (this.getWebSocketServerImplementation())(this);
+    this.webSocketServer = new (this.getServerTransport())(this);
     this.webSocketServer.implementation.on("connection", (client, request) => {
       const headers =
         // eslint-disable-next-line no-nested-ternary
@@ -1192,7 +1603,6 @@ class Server {
   }
 
   openBrowser(defaultOpenTarget) {
-    const isAbsoluteUrl = require("is-absolute-url");
     const open = require("open");
 
     Promise.all(
@@ -1202,7 +1612,7 @@ class Server {
         if (item.target === "<url>") {
           openTarget = defaultOpenTarget;
         } else {
-          openTarget = isAbsoluteUrl(item.target)
+          openTarget = Server.isAbsoluteURL(item.target)
             ? item.target
             : new URL(item.target, defaultOpenTarget).toString();
         }
@@ -1248,13 +1658,16 @@ class Server {
   logStatus() {
     const colorette = require("colorette");
 
-    const getColorsOption = (configArray) => {
-      const statsOption = this.getStatsOption(configArray);
+    const getColorsOption = (compilerOptions) => {
+      let colorsEnabled;
 
-      let colorsEnabled = false;
-
-      if (typeof statsOption === "object" && statsOption.colors) {
-        colorsEnabled = statsOption.colors;
+      if (
+        compilerOptions.stats &&
+        typeof compilerOptions.stats.colors !== "undefined"
+      ) {
+        colorsEnabled = compilerOptions.stats;
+      } else {
+        colorsEnabled = colorette.options.enabled;
       }
 
       return colorsEnabled;
@@ -1276,7 +1689,7 @@ class Server {
         return msg;
       },
     };
-    const useColor = getColorsOption(this.getCompilerConfigArray());
+    const useColor = getColorsOption(this.getCompilerOptions());
 
     if (this.options.ipc) {
       this.logger.info(`Project is running at: "${this.server.address()}"`);
@@ -1406,184 +1819,6 @@ class Server {
         `Broadcasting "${bonjourProtocol}" with subtype of "webpack" via ZeroConf DNS (Bonjour)`
       );
     }
-  }
-
-  listen(port, hostname, fn) {
-    this.logger = this.compiler.getInfrastructureLogger("webpack-dev-server");
-    this.normalizeOptions(this.options);
-
-    if (typeof port === "function") {
-      fn = port;
-    }
-
-    if (
-      typeof port !== "undefined" &&
-      typeof this.options.port !== "undefined" &&
-      port !== this.options.port
-    ) {
-      this.options.port = port;
-
-      this.logger.warn(
-        'The "port" specified in options is different from the port passed as an argument. Will be used from arguments.'
-      );
-    }
-
-    if (!this.options.port) {
-      this.options.port = port;
-    }
-
-    if (
-      typeof hostname !== "undefined" &&
-      typeof this.options.host !== "undefined" &&
-      hostname !== this.options.host
-    ) {
-      this.options.host = hostname;
-
-      this.logger.warn(
-        'The "host" specified in options is different from the host passed as an argument. Will be used from arguments.'
-      );
-    }
-
-    if (!this.options.host) {
-      this.options.host = hostname;
-    }
-
-    this.options.host = Server.getHostname(this.options.host);
-
-    const resolveFreePortOrIPC = () => {
-      if (this.options.ipc) {
-        return new Promise((resolve, reject) => {
-          const net = require("net");
-          const socket = new net.Socket();
-
-          socket.on("error", (error) => {
-            if (error.code === "ECONNREFUSED") {
-              fs.unlinkSync(this.options.ipc);
-
-              resolve(this.options.ipc);
-
-              return;
-            } else if (error.code === "ENOENT") {
-              resolve(this.options.ipc);
-
-              return;
-            }
-
-            reject(error);
-          });
-
-          socket.connect({ path: this.options.ipc }, () => {
-            throw new Error(`IPC "${this.options.ipc}" is already used`);
-          });
-        });
-      }
-
-      return Server.getFreePort(this.options.port).then((foundPort) => {
-        this.options.port = foundPort;
-      });
-    };
-
-    return resolveFreePortOrIPC()
-      .then(() => {
-        this.initialize();
-
-        const listenOptions = this.options.ipc
-          ? { path: this.options.ipc }
-          : {
-              host: this.options.host,
-              port: this.options.port,
-            };
-
-        return this.server.listen(listenOptions, (error) => {
-          if (this.options.ipc) {
-            // chmod 666 (rw rw rw)
-            const READ_WRITE = 438;
-
-            fs.chmodSync(this.options.ipc, READ_WRITE);
-          }
-
-          if (this.options.webSocketServer) {
-            try {
-              this.createWebSocketServer();
-            } catch (webSocketServerError) {
-              fn.call(this.server, webSocketServerError);
-
-              return;
-            }
-          }
-
-          if (this.options.bonjour) {
-            this.runBonjour();
-          }
-
-          this.logStatus();
-
-          if (fn) {
-            fn.call(this.server, error);
-          }
-
-          if (typeof this.options.onListening === "function") {
-            this.options.onListening(this);
-          }
-        });
-      })
-      .catch((error) => {
-        if (fn) {
-          fn.call(this.server, error);
-        }
-      });
-  }
-
-  close(callback) {
-    if (this.webSocketServer) {
-      this.webSocketServer.implementation.close();
-    }
-
-    const prom = Promise.all(
-      this.staticWatchers.map((watcher) => watcher.close())
-    );
-    this.staticWatchers = [];
-
-    if (this.server) {
-      this.server.kill(() => {
-        // watchers must be closed before closing middleware
-        prom.then(() => {
-          this.middleware.close(callback);
-        });
-      });
-    } else if (callback) {
-      callback();
-    }
-  }
-
-  // eslint-disable-next-line class-methods-use-this
-  getStatsOption(configArray) {
-    const isEmptyObject = (val) =>
-      typeof val === "object" && Object.keys(val).length === 0;
-
-    // in webpack@4 stats will not be defined if not provided,
-    // but in webpack@5 it will be an empty object
-    const statsConfig = configArray.find(
-      (configuration) =>
-        typeof configuration === "object" &&
-        configuration.stats &&
-        !isEmptyObject(configuration.stats)
-    );
-
-    return statsConfig ? statsConfig.stats : {};
-  }
-
-  getStats(statsObj) {
-    const stats = Server.DEFAULT_STATS;
-
-    const configArray = this.getCompilerConfigArray(this.compiler);
-    const statsOption = this.getStatsOption(configArray);
-
-    if (typeof statsOption === "object" && statsOption.warningsFilter) {
-      stats.warningsFilter = statsOption.warningsFilter;
-    }
-
-    return statsObj.toJson(stats);
   }
 
   setHeaders(req, res, next) {
@@ -1804,6 +2039,219 @@ class Server {
     if (this.middleware) {
       this.middleware.invalidate(callback);
     }
+  }
+
+  async start() {
+    await this.normalizeOptions();
+
+    if (this.options.ipc) {
+      await new Promise((resolve, reject) => {
+        const net = require("net");
+        const socket = new net.Socket();
+
+        socket.on("error", (error) => {
+          if (error.code === "ECONNREFUSED") {
+            // No other server listening on this socket so it can be safely removed
+            fs.unlinkSync(this.options.ipc);
+
+            resolve();
+
+            return;
+          } else if (error.code === "ENOENT") {
+            resolve();
+
+            return;
+          }
+
+          reject(error);
+        });
+
+        socket.connect({ path: this.options.ipc }, () => {
+          throw new Error(`IPC "${this.options.ipc}" is already used`);
+        });
+      });
+    } else {
+      this.options.host = await Server.getHostname(this.options.host);
+      this.options.port = await Server.getFreePort(this.options.port);
+    }
+
+    await this.initialize();
+
+    const listenOptions = this.options.ipc
+      ? { path: this.options.ipc }
+      : { host: this.options.host, port: this.options.port };
+
+    await new Promise((resolve) => {
+      this.server.listen(listenOptions, () => {
+        resolve();
+      });
+    });
+
+    if (this.options.ipc) {
+      // chmod 666 (rw rw rw)
+      const READ_WRITE = 438;
+
+      await fs.promises.chmod(this.options.ipc, READ_WRITE);
+    }
+
+    if (this.options.webSocketServer) {
+      this.createWebSocketServer();
+    }
+
+    if (this.options.bonjour) {
+      this.runBonjour();
+    }
+
+    this.logStatus();
+
+    if (typeof this.options.onListening === "function") {
+      this.options.onListening(this);
+    }
+  }
+
+  startCallback(callback) {
+    this.start().then(() => callback(null), callback);
+  }
+
+  async stop() {
+    this.webSocketProxies = [];
+
+    await Promise.all(this.staticWatchers.map((watcher) => watcher.close()));
+
+    this.staticWatchers = [];
+
+    if (this.webSocketServer) {
+      await new Promise((resolve) => {
+        this.webSocketServer.implementation.close(() => {
+          this.webSocketServer = null;
+
+          resolve();
+        });
+
+        for (const client of this.webSocketServer.clients) {
+          client.terminate();
+        }
+
+        this.webSocketServer.clients = [];
+      });
+    }
+
+    if (this.server) {
+      await new Promise((resolve) => {
+        this.server.close(() => {
+          this.server = null;
+
+          resolve();
+        });
+
+        for (const socket of this.sockets) {
+          socket.destroy();
+        }
+
+        this.sockets = [];
+      });
+
+      if (this.middleware) {
+        await new Promise((resolve, reject) => {
+          this.middleware.close((error) => {
+            if (error) {
+              reject(error);
+
+              return;
+            }
+
+            resolve();
+          });
+        });
+
+        this.middleware = null;
+      }
+    }
+  }
+
+  stopCallback(callback) {
+    this.stop().then(() => callback(null), callback);
+  }
+
+  // TODO remove in the next major release
+  listen(port, hostname, fn) {
+    util.deprecate(
+      () => {},
+      "'listen' is deprecated. Please use async 'start' or 'startCallback' methods.",
+      "DEP_WEBPACK_DEV_SERVER_LISTEN"
+    )();
+
+    this.logger = this.compiler.getInfrastructureLogger("webpack-dev-server");
+
+    if (typeof port === "function") {
+      fn = port;
+    }
+
+    if (
+      typeof port !== "undefined" &&
+      typeof this.options.port !== "undefined" &&
+      port !== this.options.port
+    ) {
+      this.options.port = port;
+
+      this.logger.warn(
+        'The "port" specified in options is different from the port passed as an argument. Will be used from arguments.'
+      );
+    }
+
+    if (!this.options.port) {
+      this.options.port = port;
+    }
+
+    if (
+      typeof hostname !== "undefined" &&
+      typeof this.options.host !== "undefined" &&
+      hostname !== this.options.host
+    ) {
+      this.options.host = hostname;
+
+      this.logger.warn(
+        'The "host" specified in options is different from the host passed as an argument. Will be used from arguments.'
+      );
+    }
+
+    if (!this.options.host) {
+      this.options.host = hostname;
+    }
+
+    return this.start()
+      .then(() => {
+        if (fn) {
+          fn.call(this.server);
+        }
+      })
+      .catch((error) => {
+        // Nothing
+        if (fn) {
+          fn.call(this.server, error);
+        }
+      });
+  }
+
+  // TODO remove in the next major release
+  close(callback) {
+    util.deprecate(
+      () => {},
+      "'close' is deprecated. Please use async 'stop' or 'stopCallback' methods.",
+      "DEP_WEBPACK_DEV_SERVER_CLOSE"
+    )();
+
+    return this.stop()
+      .then(() => {
+        if (callback) {
+          callback(null);
+        }
+      })
+      .catch((error) => {
+        if (callback) {
+          callback(error);
+        }
+      });
   }
 }
 
