@@ -5,6 +5,6 @@
 module.exports = class BaseServer {
   constructor(server) {
     this.server = server;
-    this.clients = new Set();
+    this.clients = [];
   }
 };
