@@ -67,19 +67,15 @@ module.exports = class SockJSServer extends BaseServer {
       client.send = client.write;
       client.terminate = client.close;
 
-      this.clients.add(client);
+      this.clients.push(client);
 
       client.on("close", () => {
-        this.clients.delete(client);
+        this.clients.splice(this.clients.indexOf(client), 1);
       });
     });
 
-    this.implementation.close = () => {
-      for (const client of this.clients) {
-        client.close();
-      }
-
-      this.clients.clear();
+    this.implementation.close = (callback) => {
+      callback();
     };
   }
 };
