@@ -54,7 +54,7 @@ module.exports = class WebsocketServer extends BaseServer {
     }, WebsocketServer.heartbeatInterval);
 
     this.implementation.on("connection", (client) => {
-      this.clients.add(client);
+      this.clients.push(client);
 
       client.isAlive = true;
 
@@ -63,14 +63,12 @@ module.exports = class WebsocketServer extends BaseServer {
       });
 
       client.on("close", () => {
-        this.clients.delete(client);
+        this.clients.splice(this.clients.indexOf(client), 1);
       });
     });
 
     this.implementation.on("close", () => {
       clearInterval(interval);
-
-      this.clients.clear();
     });
   }
 };
