@@ -43,22 +43,11 @@ describe("entry", () => {
   it("should work with single entry", async () => {
     const compiler = webpack({ ...config, entry: entryFirst });
     const devServerOptions = {
-      host: "127.0.0.1",
       port,
     };
     const server = new Server(devServerOptions, compiler);
 
-    await new Promise((resolve, reject) => {
-      server.listen(port, "127.0.0.1", (error) => {
-        if (error) {
-          reject(error);
-
-          return;
-        }
-
-        resolve();
-      });
-    });
+    await server.start();
 
     const { page, browser } = await runBrowser();
 
@@ -83,38 +72,17 @@ describe("entry", () => {
     expect(pageErrors).toMatchSnapshot("page errors");
 
     await browser.close();
-    await new Promise((resolve, reject) => {
-      server.close((error) => {
-        if (error) {
-          reject(error);
-
-          return;
-        }
-
-        resolve();
-      });
-    });
+    await server.stop();
   });
 
   it("should work with single array entry", async () => {
     const compiler = webpack({ ...config, entry: [entryFirst, entrySecond] });
     const devServerOptions = {
-      host: "127.0.0.1",
       port,
     };
     const server = new Server(devServerOptions, compiler);
 
-    await new Promise((resolve, reject) => {
-      server.listen(port, "127.0.0.1", (error) => {
-        if (error) {
-          reject(error);
-
-          return;
-        }
-
-        resolve();
-      });
-    });
+    await server.start();
 
     const { page, browser } = await runBrowser();
 
@@ -139,17 +107,7 @@ describe("entry", () => {
     expect(pageErrors).toMatchSnapshot("page errors");
 
     await browser.close();
-    await new Promise((resolve, reject) => {
-      server.close((error) => {
-        if (error) {
-          reject(error);
-
-          return;
-        }
-
-        resolve();
-      });
-    });
+    await server.stop();
   });
 
   itOnlyWebpack5("should work with object entry", async () => {
@@ -160,22 +118,11 @@ describe("entry", () => {
       },
     });
     const devServerOptions = {
-      host: "127.0.0.1",
       port,
     };
     const server = new Server(devServerOptions, compiler);
 
-    await new Promise((resolve, reject) => {
-      server.listen(port, "127.0.0.1", (error) => {
-        if (error) {
-          reject(error);
-
-          return;
-        }
-
-        resolve();
-      });
-    });
+    await server.start();
 
     const { page, browser } = await runBrowser();
 
@@ -200,38 +147,17 @@ describe("entry", () => {
     expect(pageErrors).toMatchSnapshot("page errors");
 
     await browser.close();
-    await new Promise((resolve, reject) => {
-      server.close((error) => {
-        if (error) {
-          reject(error);
-
-          return;
-        }
-
-        resolve();
-      });
-    });
+    await server.stop();
   });
 
   it("should work with dynamic entry", async () => {
     const compiler = webpack({ ...config, entry: () => entryFirst });
     const devServerOptions = {
-      host: "127.0.0.1",
       port,
     };
     const server = new Server(devServerOptions, compiler);
 
-    await new Promise((resolve, reject) => {
-      server.listen(port, "127.0.0.1", (error) => {
-        if (error) {
-          reject(error);
-
-          return;
-        }
-
-        resolve();
-      });
-    });
+    await server.start();
 
     const { page, browser } = await runBrowser();
 
@@ -256,17 +182,7 @@ describe("entry", () => {
     expect(pageErrors).toMatchSnapshot("page errors");
 
     await browser.close();
-    await new Promise((resolve, reject) => {
-      server.close((error) => {
-        if (error) {
-          reject(error);
-
-          return;
-        }
-
-        resolve();
-      });
-    });
+    await server.stop();
   });
 
   it("should work with dynamic async entry", async () => {
@@ -275,22 +191,11 @@ describe("entry", () => {
       entry: () => new Promise((resolve) => resolve([entryFirst])),
     });
     const devServerOptions = {
-      host: "127.0.0.1",
       port,
     };
     const server = new Server(devServerOptions, compiler);
 
-    await new Promise((resolve, reject) => {
-      server.listen(port, "127.0.0.1", (error) => {
-        if (error) {
-          reject(error);
-
-          return;
-        }
-
-        resolve();
-      });
-    });
+    await server.start();
 
     const { page, browser } = await runBrowser();
 
@@ -315,17 +220,7 @@ describe("entry", () => {
     expect(pageErrors).toMatchSnapshot("page errors");
 
     await browser.close();
-    await new Promise((resolve, reject) => {
-      server.close((error) => {
-        if (error) {
-          reject(error);
-
-          return;
-        }
-
-        resolve();
-      });
-    });
+    await server.stop();
   });
 
   it("should work with multiple entries", async () => {
@@ -342,22 +237,11 @@ describe("entry", () => {
       },
     });
     const devServerOptions = {
-      host: "127.0.0.1",
       port,
     };
     const server = new Server(devServerOptions, compiler);
 
-    await new Promise((resolve, reject) => {
-      server.listen(port, "127.0.0.1", (error) => {
-        if (error) {
-          reject(error);
-
-          return;
-        }
-
-        resolve();
-      });
-    });
+    await server.start();
 
     const { page, browser } = await runBrowser();
 
@@ -383,17 +267,7 @@ describe("entry", () => {
     expect(pageErrors).toMatchSnapshot("page errors");
 
     await browser.close();
-    await new Promise((resolve, reject) => {
-      server.close((error) => {
-        if (error) {
-          reject(error);
-
-          return;
-        }
-
-        resolve();
-      });
-    });
+    await server.stop();
   });
 
   it("should work with multiple entries #2", async () => {
@@ -410,22 +284,11 @@ describe("entry", () => {
       },
     });
     const devServerOptions = {
-      host: "127.0.0.1",
       port,
     };
     const server = new Server(devServerOptions, compiler);
 
-    await new Promise((resolve, reject) => {
-      server.listen(port, "127.0.0.1", (error) => {
-        if (error) {
-          reject(error);
-
-          return;
-        }
-
-        resolve();
-      });
-    });
+    await server.start();
 
     const { page, browser } = await runBrowser();
     const pageErrors = [];
@@ -450,17 +313,7 @@ describe("entry", () => {
     expect(pageErrors).toMatchSnapshot("page errors");
 
     await browser.close();
-    await new Promise((resolve, reject) => {
-      server.close((error) => {
-        if (error) {
-          reject(error);
-
-          return;
-        }
-
-        resolve();
-      });
-    });
+    await server.stop();
   });
 
   itOnlyWebpack5(
@@ -477,22 +330,11 @@ describe("entry", () => {
         },
       });
       const devServerOptions = {
-        host: "127.0.0.1",
         port,
       };
       const server = new Server(devServerOptions, compiler);
 
-      await new Promise((resolve, reject) => {
-        server.listen(port, "127.0.0.1", (error) => {
-          if (error) {
-            reject(error);
-
-            return;
-          }
-
-          resolve();
-        });
-      });
+      await server.start();
 
       const { page, browser } = await runBrowser();
 
@@ -518,17 +360,7 @@ describe("entry", () => {
       expect(pageErrors).toMatchSnapshot("page errors");
 
       await browser.close();
-      await new Promise((resolve, reject) => {
-        server.close((error) => {
-          if (error) {
-            reject(error);
-
-            return;
-          }
-
-          resolve();
-        });
-      });
+      await server.stop();
     }
   );
 
@@ -544,22 +376,11 @@ describe("entry", () => {
     }).apply(compiler);
 
     const devServerOptions = {
-      host: "127.0.0.1",
       port,
     };
     const server = new Server(devServerOptions, compiler);
 
-    await new Promise((resolve, reject) => {
-      server.listen(port, "127.0.0.1", (error) => {
-        if (error) {
-          reject(error);
-
-          return;
-        }
-
-        resolve();
-      });
-    });
+    await server.start();
 
     const { page, browser } = await runBrowser();
 
@@ -584,16 +405,6 @@ describe("entry", () => {
     expect(pageErrors).toMatchSnapshot("page errors");
 
     await browser.close();
-    await new Promise((resolve, reject) => {
-      server.close((error) => {
-        if (error) {
-          reject(error);
-
-          return;
-        }
-
-        resolve();
-      });
-    });
+    await server.stop();
   });
 });
