@@ -20,7 +20,6 @@ describe("progress", () => {
     const compiler = webpack(reloadConfig);
     const devServerOptions = {
       port,
-      host: "0.0.0.0",
       static: false,
       hot: true,
       client: {
@@ -29,17 +28,7 @@ describe("progress", () => {
     };
     const server = new Server(devServerOptions, compiler);
 
-    await new Promise((resolve, reject) => {
-      server.listen(devServerOptions.port, devServerOptions.host, (error) => {
-        if (error) {
-          reject(error);
-
-          return;
-        }
-
-        resolve();
-      });
-    });
+    await server.start();
 
     const { page, browser } = await runBrowser();
 
@@ -85,10 +74,6 @@ describe("progress", () => {
 
     fs.unlinkSync(cssFilePath);
 
-    await new Promise((resolve) => {
-      server.close(() => {
-        resolve();
-      });
-    });
+    await server.stop();
   });
 });
