@@ -6,5 +6,5 @@ module.exports = {
   mode: "development",
   context: __dirname,
   entry: "./foo.js",
-  plugins: [ExitOnDonePlugin],
+  plugins: [new ExitOnDonePlugin()],
 };
