@@ -1,5 +1,9 @@
 "use strict";
 
+const webpack = require("webpack");
+
+const isWebpack5 = webpack.version.startsWith("5");
+
 const HTMLContent = `
 <!doctype html>
 <html>
@@ -20,9 +24,16 @@ module.exports = {
   output: {
     path: "/",
   },
-  infrastructureLogging: {
-    level: "warn",
-  },
+  infrastructureLogging: isWebpack5
+    ? {
+        level: "info",
+        stream: {
+          write: () => {},
+        },
+      }
+    : {
+        level: "info",
+      },
   plugins: [
     {
       apply(compiler) {
