@@ -1,5 +1,9 @@
 "use strict";
 
+const webpack = require("webpack");
+
+const isWebpack5 = webpack.version.startsWith("5");
+
 module.exports = [
   {
     mode: "development",
@@ -10,8 +14,15 @@ module.exports = [
       path: "/",
     },
     node: false,
-    infrastructureLogging: {
-      level: "warn",
-    },
+    infrastructureLogging: isWebpack5
+      ? {
+          level: "info",
+          stream: {
+            write: () => {},
+          },
+        }
+      : {
+          level: "info",
+        },
   },
 ];
