@@ -1,5 +1,9 @@
 "use strict";
 
+const webpack = require("webpack");
+
+const isWebpack5 = webpack.version.startsWith("5");
+
 module.exports = {
   mode: "development",
   context: __dirname,
@@ -16,8 +20,14 @@ module.exports = {
       },
     ],
   },
-  infrastructureLogging: {
-    level: "warn",
-  },
-  node: false,
+  infrastructureLogging: isWebpack5
+    ? {
+        level: "info",
+        stream: {
+          write: () => {},
+        },
+      }
+    : {
+        level: "info",
+      },
 };
