@@ -20,6 +20,8 @@ describe('"port" option', () => {
     beforeAll(async () => {
       const compiler = webpack(config);
 
+      process.env.WEBPACK_DEV_SERVER_BASE_PORT = 10020;
+
       server = new Server(
         {
           port,
@@ -31,33 +33,21 @@ describe('"port" option', () => {
         compiler
       );
 
-      await new Promise((resolve, reject) => {
-        server.listen(port, "127.0.0.1", (error) => {
-          if (error) {
-            reject(error);
-
-            return;
-          }
-
-          resolve();
-        });
-      });
+      await server.start();
 
       req = request(server.app);
     });
 
     afterAll(async () => {
-      await new Promise((resolve) => {
-        server.close(() => {
-          resolve();
-        });
-      });
+      delete process.env.WEBPACK_DEV_SERVER_BASE_PORT;
+
+      await server.stop();
     });
 
     it("server address", () => {
       const address = server.server.address();
 
-      expect(address.address).toBe("127.0.0.1");
+      expect(address.address).toBe("::");
       // Random port
       expect(address.port).toBeDefined();
     });
@@ -73,6 +63,8 @@ describe('"port" option', () => {
     beforeAll(async () => {
       const compiler = webpack(config);
 
+      process.env.WEBPACK_DEV_SERVER_BASE_PORT = 10030;
+
       server = new Server(
         {
           // eslint-disable-next-line no-undefined
@@ -85,33 +77,21 @@ describe('"port" option', () => {
         compiler
       );
 
-      await new Promise((resolve, reject) => {
-        server.listen(port, "127.0.0.1", (error) => {
-          if (error) {
-            reject(error);
-
-            return;
-          }
-
-          resolve();
-        });
-      });
+      await server.start();
 
       req = request(server.app);
     });
 
     afterAll(async () => {
-      await new Promise((resolve) => {
-        server.close(() => {
-          resolve();
-        });
-      });
+      delete process.env.WEBPACK_DEV_SERVER_BASE_PORT;
+
+      await server.stop();
     });
 
     it("server address", () => {
       const address = server.server.address();
 
-      expect(address.address).toBe("127.0.0.1");
+      expect(address.address).toBe("::");
       // Random port
       expect(address.port).toBeDefined();
     });
@@ -127,6 +107,8 @@ describe('"port" option', () => {
     beforeAll(async () => {
       const compiler = webpack(config);
 
+      process.env.WEBPACK_DEV_SERVER_BASE_PORT = 10040;
+
       server = new Server(
         {
           port: "auto",
@@ -138,33 +120,21 @@ describe('"port" option', () => {
         compiler
       );
 
-      await new Promise((resolve, reject) => {
-        server.listen(port, "127.0.0.1", (error) => {
-          if (error) {
-            reject(error);
-
-            return;
-          }
-
-          resolve();
-        });
-      });
+      await server.start();
 
       req = request(server.app);
     });
 
     afterAll(async () => {
-      await new Promise((resolve) => {
-        server.close(() => {
-          resolve();
-        });
-      });
+      delete process.env.WEBPACK_DEV_SERVER_BASE_PORT;
+
+      await server.stop();
     });
 
     it("server address", () => {
       const address = server.server.address();
 
-      expect(address.address).toBe("127.0.0.1");
+      expect(address.address).toBe("::");
       // Random port
       expect(address.port).toBeDefined();
     });
@@ -191,33 +161,19 @@ describe('"port" option', () => {
         compiler
       );
 
-      await new Promise((resolve, reject) => {
-        server.listen("33333", "127.0.0.1", (error) => {
-          if (error) {
-            reject(error);
-
-            return;
-          }
-
-          resolve();
-        });
-      });
+      await server.start();
 
       req = request(server.app);
     });
 
     afterAll(async () => {
-      await new Promise((resolve) => {
-        server.close(() => {
-          resolve();
-        });
-      });
+      await server.stop();
     });
 
     it("server address", () => {
       const address = server.server.address();
 
-      expect(address.address).toBe("127.0.0.1");
+      expect(address.address).toBe("::");
       expect(address.port).toBe(33333);
     });
 
@@ -243,33 +199,19 @@ describe('"port" option', () => {
         compiler
       );
 
-      await new Promise((resolve, reject) => {
-        server.listen(33333, "127.0.0.1", (error) => {
-          if (error) {
-            reject(error);
-
-            return;
-          }
-
-          resolve();
-        });
-      });
+      await server.start();
 
       req = request(server.app);
     });
 
     afterAll(async () => {
-      await new Promise((resolve) => {
-        server.close(() => {
-          resolve();
-        });
-      });
+      await server.stop();
     });
 
     it("server address", () => {
       const address = server.server.address();
 
-      expect(address.address).toBe("127.0.0.1");
+      expect(address.address).toBe("::");
       expect(address.port).toBe(33333);
     });
 
