@@ -7,9 +7,11 @@ const port = require("../ports-map")["setup-exit-signals-option"];
 
 describe("setupExitSignals option", () => {
   let server;
+  let doExit;
   let exitSpy;
-  let killSpy;
+  let stopCallbackSpy;
   let stdinResumeSpy;
+
   const signals = ["SIGINT", "SIGTERM"];
 
   beforeEach(async () => {
@@ -23,23 +25,17 @@ describe("setupExitSignals option", () => {
       compiler
     );
 
-    await new Promise((resolve, reject) => {
-      server.listen(port, "127.0.0.1", (error) => {
-        if (error) {
-          reject(error);
+    await server.start();
 
-          return;
-        }
+    doExit = false;
 
-        resolve();
-      });
+    exitSpy = jest.spyOn(process, "exit").mockImplementation(() => {
+      doExit = true;
     });
-
-    exitSpy = jest.spyOn(process, "exit").mockImplementation(() => {});
     stdinResumeSpy = jest
       .spyOn(process.stdin, "resume")
       .mockImplementation(() => {});
-    killSpy = jest.spyOn(server.server, "kill");
+    stopCallbackSpy = jest.spyOn(server, "stopCallback");
   });
 
   afterEach(async () => {
@@ -50,19 +46,22 @@ describe("setupExitSignals option", () => {
     });
     process.stdin.removeAllListeners("end");
 
-    await new Promise((resolve) => {
-      server.close(() => {
-        resolve();
-      });
-    });
+    await server.stop();
   });
 
-  it.each(signals)("should close and exit on %s", (signal, done) => {
+  it.each(signals)("should close and exit on %s", async (signal) => {
     process.emit(signal);
-    process.nextTick(() => {
-      expect(killSpy.mock.calls.length).toEqual(1);
 
-      done();
+    await new Promise((resolve) => {
+      const interval = setInterval(() => {
+        if (doExit) {
+          expect(stopCallbackSpy.mock.calls.length).toEqual(1);
+
+          clearInterval(interval);
+
+          resolve();
+        }
+      }, 100);
     });
   });
 });
