@@ -37,27 +37,13 @@ describe("static.directory option", () => {
         compiler
       );
 
-      await new Promise((resolve, reject) => {
-        server.listen(port, "127.0.0.1", (error) => {
-          if (error) {
-            reject(error);
-
-            return;
-          }
-
-          resolve();
-        });
-      });
+      await server.start();
 
       req = request(server.app);
     });
 
     afterAll(async () => {
-      await new Promise((resolve) => {
-        server.close(() => {
-          resolve();
-        });
-      });
+      await server.stop();
 
       fs.truncateSync(nestedFile);
     });
@@ -125,27 +111,13 @@ describe("static.directory option", () => {
         compiler
       );
 
-      await new Promise((resolve, reject) => {
-        server.listen(port, "127.0.0.1", (error) => {
-          if (error) {
-            reject(error);
-
-            return;
-          }
-
-          resolve();
-        });
-      });
+      await server.start();
 
       req = request(server.app);
     });
 
     afterAll(async () => {
-      await new Promise((resolve) => {
-        server.close(() => {
-          resolve();
-        });
-      });
+      await server.stop();
     });
 
     it("shouldn't list the files inside the assets folder (404)", async () => {
@@ -178,27 +150,13 @@ describe("static.directory option", () => {
         compiler
       );
 
-      await new Promise((resolve, reject) => {
-        server.listen(port, "127.0.0.1", (error) => {
-          if (error) {
-            reject(error);
-
-            return;
-          }
-
-          resolve();
-        });
-      });
+      await server.start();
 
       req = request(server.app);
     });
 
     afterAll(async () => {
-      await new Promise((resolve) => {
-        server.close(() => {
-          resolve();
-        });
-      });
+      await server.stop();
     });
 
     it("should list the files inside the assets folder (200)", async () => {
@@ -230,27 +188,13 @@ describe("static.directory option", () => {
         compiler
       );
 
-      await new Promise((resolve, reject) => {
-        server.listen(port, "127.0.0.1", (error) => {
-          if (error) {
-            reject(error);
-
-            return;
-          }
-
-          resolve();
-        });
-      });
+      await server.start();
 
       req = request(server.app);
     });
 
     afterAll(async () => {
-      await new Promise((resolve) => {
-        server.close(() => {
-          resolve();
-        });
-      });
+      await server.stop();
     });
 
     it("should list the files inside the assets folder (200)", async () => {
@@ -279,27 +223,13 @@ describe("static.directory option", () => {
         compiler
       );
 
-      await new Promise((resolve, reject) => {
-        server.listen(port, "127.0.0.1", (error) => {
-          if (error) {
-            reject(error);
-
-            return;
-          }
-
-          resolve();
-        });
-      });
+      await server.start();
 
       req = request(server.app);
     });
 
     afterAll(async () => {
-      await new Promise((resolve) => {
-        server.close(() => {
-          resolve();
-        });
-      });
+      await server.stop();
     });
 
     it("Request to first directory", async () => {
@@ -325,19 +255,22 @@ describe("static.directory option", () => {
     });
 
     it("Should throw exception (external url)", (done) => {
-      try {
-        // eslint-disable-next-line no-unused-vars
-        server = testServer.start(config, {
-          static: "https://example.com/",
-        });
+      expect.assertions(1);
 
-        expect(true).toBe(false);
-      } catch (e) {
-        expect(e.message).toBe(
-          "Using a URL as static.directory is not supported"
-        );
-        done();
-      }
+      // eslint-disable-next-line no-unused-vars
+      server = testServer.start(
+        config,
+        {
+          static: "https://example.com/",
+        },
+        (error) => {
+          expect(error.message).toBe(
+            "Using a URL as static.directory is not supported"
+          );
+
+          server.close(done);
+        }
+      );
     });
 
     it("Should not throw exception (local path with lower case first character)", (done) => {
@@ -385,19 +318,20 @@ describe("static.directory option", () => {
     });
 
     it("Should throw exception (array with absolute url)", (done) => {
-      try {
-        // eslint-disable-next-line no-unused-vars
-        server = testServer.start(config, {
+      // eslint-disable-next-line no-unused-vars
+      server = testServer.start(
+        config,
+        {
           static: [publicDirectory, "https://example.com/"],
-        });
+        },
+        (error) => {
+          expect(error.message).toBe(
+            "Using a URL as static.directory is not supported"
+          );
 
-        expect(true).toBe(false);
-      } catch (e) {
-        expect(e.message).toBe(
-          "Using a URL as static.directory is not supported"
-        );
-        done();
-      }
+          server.close(done);
+        }
+      );
     });
   });
 
@@ -418,27 +352,13 @@ describe("static.directory option", () => {
         compiler
       );
 
-      await new Promise((resolve, reject) => {
-        server.listen(port, "127.0.0.1", (error) => {
-          if (error) {
-            reject(error);
-
-            return;
-          }
-
-          resolve();
-        });
-      });
+      await server.start();
 
       req = request(server.app);
     });
 
     afterAll(async () => {
-      await new Promise((resolve) => {
-        server.close(() => {
-          resolve();
-        });
-      });
+      await server.stop();
     });
 
     it("Request to page", async () => {
@@ -464,27 +384,13 @@ describe("static.directory option", () => {
         compiler
       );
 
-      await new Promise((resolve, reject) => {
-        server.listen(port, "127.0.0.1", (error) => {
-          if (error) {
-            reject(error);
-
-            return;
-          }
-
-          resolve();
-        });
-      });
+      await server.start();
 
       req = request(server.app);
     });
 
     afterAll(async () => {
-      await new Promise((resolve) => {
-        server.close(() => {
-          resolve();
-        });
-      });
+      await server.stop();
     });
 
     it("Request to page", async () => {
